refactor(frontend): split CustomSubdivisionDialog AI controls into hooks + components
Extract AI assistance logic from CustomSubdivisionDialog:
- AIAssistControls.tsx: rendered AI controls bar (model select, generate button,
  usage popover trigger)
- AIDivisionList.tsx: AI-suggested division list with accept/reject per item
- useAISuggestions.ts: hook managing suggestion fetch, loading state, and
  apply/reject callbacks
- useGroupDescriptions.ts: hook for generating and caching group descriptions
  via the AI API

Update index.tsx and AIAssistTab.tsx to consume the new components/hooks.
